requirements.txt
slackclient~=2.9.3
requests~=2.26.0
boto3~=1.18.6
Unidecode~=1.2.0
pendulum~=2.1.2
expiringdict~=1.2.1
psycopg2-binary~=2.9.1
